Fix compile errors on NetBSD-current (gcc 4.1.3).
git-svn-id: svn+ssh://combine-101/home/svn_bio/trunk@140 450597fd-3c47-0410-84d8-f85f2794411d

+++ splittigs/splittigs.c
@@ -14,11 +14,11 @@ strtrim(char *str)
 	if (str == NULL)
 		return (NULL);
 
-	while (isspace(*str))
+	while (isspace((int)*str))
 		str++;
 
 	for (i = strlen(str) - 1; i >= 0; i--) {
-		if (!isspace(str[i]))
+		if (!isspace((int)str[i]))
 			break;
 		str[i] = '\0';
 	}
